Move library redux down one level
